ZTools/.trae/skills/zentao-helper/requirements.txt
# ZenTao Helper Skill 依赖

# HTTP 请求
requests>=2.31.0
urllib3>=2.0.0

# 配置管理
PyYAML>=6.0.0

# 加密存储
cryptography>=41.0.0
keyring>=24.0.0

# 日志（使用 Python 标准库，无需额外依赖）